+++ module/data/npc-base-data.mjs
@@ -39,7 +39,7 @@ export function npcBaseFields() {
       validationError: game.i18n.localize('SDM.ErrorValidationRollFormula')
     }),
 
-    capacity: new fields.NumberField({ ...requiredInteger, initial: 1, min: 1 })
+    capacity: new fields.NumberField({ required: true, initial: 1, min: 0 })
   };
 }